[iOS] 중복된 빌드타겟 파일들 정리
