+++ ios/meditation_blossom/SermonBuilder.m
@@ -5,20 +5,23 @@
 //  Created by 최상준 on 4/9/26.
 //
 
-#import "sermonBuilder.h"
+#import "SermonBuilder.h"
 
 @implementation SermonBuilder
 
 + (NSMutableDictionary *)buildFromPayload:(NSDictionary *)data
                                  sourceId:(NSString *)sourceId
 {
+    NSArray<NSDictionary *> *parsedBibleReferences =
+        [SermonBuilder parsedBibleReferencesFromValue:data[@"bible_references"]];
+
     NSMutableDictionary *sermonData = [@{
         @"id": sourceId ?: @"",
         @"source_id": sourceId ?: @"",
         @"title": data[@"title"] ?: @"",
         @"content": data[@"content"] ?: @"",
         @"category": data[@"category"] ?: @"",
-        @"bible_references": data[@"bible_references"] ?: @"",
+        @"bible_references": parsedBibleReferences ?: @[],
         @"meditation_questions": data[@"meditation_questions"] ?: @"",
         @"date": data[@"date"] ?: @"",
         @"year": data[@"year"] ?: @"",
@@ -30,11 +33,56 @@
         @"topic": data[@"topic"] ?: @""
     } mutableCopy];
 
-    if (data[@"video_url"]) {
+    if (data[@"video_url"] && data[@"video_url"] != [NSNull null]) {
         sermonData[@"video_url"] = data[@"video_url"];
     }
 
     return sermonData;
+}
+
++ (NSArray<NSDictionary *> *)parsedBibleReferencesFromValue:(id)rawValue
+{
+    if (!rawValue || rawValue == [NSNull null]) {
+        return @[];
+    }
+
+    if ([rawValue isKindOfClass:[NSArray class]]) {
+        NSMutableArray<NSDictionary *> *result = [NSMutableArray array];
+        for (id item in (NSArray *)rawValue) {
+            if ([item isKindOfClass:[NSDictionary class]]) {
+                [result addObject:item];
+            }
+        }
+        return [result copy];
+    }
+
+    if ([rawValue isKindOfClass:[NSString class]]) {
+        NSString *jsonString = (NSString *)rawValue;
+        NSData *jsonData = [jsonString dataUsingEncoding:NSUTF8StringEncoding];
+        if (!jsonData) {
+            return @[];
+        }
+
+        NSError *error = nil;
+        id parsed = [NSJSONSerialization JSONObjectWithData:jsonData options:0 error:&error];
+        if (error) {
+            NSLog(@"❌ bible_references JSON parse error: %@", error);
+            return @[];
+        }
+
+        if ([parsed isKindOfClass:[NSArray class]]) {
+            NSMutableArray<NSDictionary *> *result = [NSMutableArray array];
+            for (id item in (NSArray *)parsed) {
+                if ([item isKindOfClass:[NSDictionary class]]) {
+                    [result addObject:item];
+                }
+            }
+            return [result copy];
+        }
+    }
+
+    NSLog(@"❌ bible_references has unsupported type: %@", NSStringFromClass([rawValue class]));
+    return @[];
 }
 
 @end
